Второе домашнее задание. Формула силы трения на наклонной поверхности. С вводом и выводом в консоль.

+++ src/main/java/lesson4/SqrtExtractingByNewtonMethod.java
@@ -21,7 +21,7 @@ public class SqrtExtractingByNewtonMethod {
 
     public static double sqrt(double n) {
         double a = 1;
-        for (double i = 1; i <= 10; i++) {
+        for (int i = 1; i <= 10; i++) {
             
             a = a - (a * a  - n) / (2 * a);
         }
